Move Event to SDK
